refactor: 파일 구조정리
