+++ src/components/auth/AuthHeader.tsx
