+++ src/components/auth/ProfileAvatar.tsx
