+++ src/pages/ProfileEditPage.tsx
@@ -1,8 +1,8 @@
 import { FormProvider } from "react-hook-form"
+import { AuthFormField } from "@/components/auth/AuthFormField"
+import { AuthHeader } from "@/components/auth/AuthHeader"
+import { ProfileAvatar } from "@/components/auth/ProfileAvatar"
 import { AlertDialog } from "@/components/common/AlertDialog"
-import { AuthFormField } from "@/components/common/AuthFormField"
-import { AuthHeader } from "@/components/common/AuthHeader"
-import { ProfileAvatar } from "@/components/common/ProfileAvatar"
 import { Button } from "@/components/ui/Button"
 import { Input } from "@/components/ui/Input"
 import { AUTH_LAYOUT, AUTH_STYLES } from "@/constants/auth-styles"
